+++ src/main/java/me/salamander/noisetest/terra/normalize/ClampNormalizerModule.java
@@ -14,7 +14,7 @@ public class ClampNormalizerModule extends NormalizerModule{
     }
 
     @Override
-    public double sample(double x, double y) {
+    public float sample(float x, float y) {
         float value = (float) function.sample(x, y);
 
         if(value < min) return min;
