+++ src/main/java/me/salamander/noisetest/terra/normalize/LinearNormalizerModule.java
@@ -17,7 +17,7 @@ public class LinearNormalizerModule extends NormalizerModule{
     }
 
     @Override
-    public double sample(double x, double y) {
+    public float sample(float x, float y) {
         return (function.sample(x, y) - min) * divisor - 1;
     }
 }
